Rationalist util.py (#200)
Also additional tidying up.
