+++ run_configs/lfric/lfric_common.py
@@ -7,7 +7,8 @@ from typing import Dict
 from fab.artefacts import SuffixFilter
 from fab.steps import Step, check_for_errors
 from fab.steps.preprocess import PreProcessor
-from fab.util import run_command, log_or_dot, input_to_output_fpath
+from fab.util import log_or_dot, input_to_output_fpath
+from fab.tools import run_command
 
 logger = logging.getLogger('fab')
