+++ source/fab/steps/grab.py
@@ -20,8 +20,7 @@ except ImportError:
     svn = None
 
 from fab.steps import Step
-from fab.util import run_command
-
+from fab.tools import run_command
 
 logger = logging.getLogger(__name__)
